✨ 모바일에서도 밈 텍스트 이동 가능하도록 구현
